feat: update Engagement and StaffProfile models, adjust data types and add managedCohorts relationship

+++ database/migrations/2026_06_06_171817_create_staff_profiles.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained('users')->cascadeOnDelete();
             $table->enum('compensation_type', ['internal', 'external']);
-            $table->decimal('hourly_rate');
+            $table->int('hourly_rate');
             $table->timestamps();
         });
     }
